Add vcxproj targeting Android and update code to support build (#139)
Update AsyncLib to support cross-compiling to Android and add a vcxproj to enable building a static library targeting Android.

+++ Include/httpClient/trace.h
@@ -163,7 +163,7 @@ STDAPI HCSettingsGetTraceLevel(
 typedef void (HCTraceCallback)(
     _In_ UTF8CSTR areaName,
     _In_ enum HCTraceLevel level,
-    _In_ uint32_t threadId,
+    _In_ uint64_t threadId,
     _In_ uint64_t timestamp,
     _In_ UTF8CSTR message
     );
